Fix test files to match source files
